Small refactoring
renamed the UserProfileFriends composable to UserProfileFriendsFinder to make it more understandable
