+++ app/src/main/java/com/example/triptracker/MainActivity.kt
@@ -30,7 +30,7 @@ import com.example.triptracker.view.profile.UserProfileEditScreen
 import com.example.triptracker.view.profile.UserProfileFavourite
 import com.example.triptracker.view.profile.UserProfileFollowers
 import com.example.triptracker.view.profile.UserProfileFollowing
-import com.example.triptracker.view.profile.UserProfileFriends
+import com.example.triptracker.view.profile.UserProfileFriendsFinder
 import com.example.triptracker.view.profile.UserProfileMyTrips
 import com.example.triptracker.view.profile.UserProfileOverview
 import com.example.triptracker.view.profile.UserProfileSettings
@@ -110,7 +110,7 @@ class MainActivity : ComponentActivity() {
                 UserProfileOverview(navigation = navigation, profile = profile)
               }
               composable(Route.FRIENDS) {
-                UserProfileFriends(navigation = navigation, profile = profile)
+                UserProfileFriendsFinder(navigation = navigation, profile = profile)
               }
               composable(Route.FOLLOWERS) {
                 UserProfileFollowers(navigation = navigation, profile = profile)
