+++ app/src/main/java/com/example/triptracker/view/profile/subviews/FriendListView.kt
@@ -1,4 +1,4 @@
-package com.example.triptracker.view.profile
+package com.example.triptracker.view.profile.subviews
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
